cloud_functions/monster_classifier/requirements.txt
functions-framework==3.*
google-cloud-bigquery==3.*
google-generativeai==0.7.*
